requirements.txt
streamlit
langchain
langchain-openai
faiss-cpu
pypdf
python-dotenv
tiktoken
langchain_community